# Conflicts:
#	Makefile
#	admin-portal/src/api/dataManager.js
#	admin-portal/src/api/globalVariable.js
#	admin-portal/src/api/modelDev.js
#	admin-portal/src/error/index.js
#	admin-portal/src/views/userManager/user.vue
#	build/application/api-doc/dockerfile
#	deploy/charts/octopus/templates/api-doc.yaml
#	deploy/charts/octopus/templates/base-server.yaml
#	deploy/charts/octopus/values.yaml
#	openai-portal/src/api/datasetManager.js
#	openai-portal/src/api/generalView.js
#	openai-portal/src/api/globalVariable.js
#	openai-portal/src/api/imageManager.js
#	openai-portal/src/api/trainingManager.js
#	server/admin-server/configs/config.yaml
#	server/base-server/configs/config.yaml
#	server/base-server/internal/conf/conf.proto
#	server/base-server/internal/data/data.go
#	server/common/errors/codes.go
#	server/common/utils/protoc.go
#	server/go.mod
#	server/go.sum
#	server/openai-server/configs/config.yaml

+++ deploy/single_master_k8s_install/node_install_k8s.sh
@@ -41,8 +41,12 @@ join_cluster() {
     kubeadm join $masterip:6443 --token $tocken --discovery-token-ca-cert-hash sha256:$sha_value --ignore-preflight-errors=…
 
     mkdir -p $HOME/.kube
-    cp -i $bash_path/config $HOME/.kube/config
-	chown $(id -u):$(id -g) $HOME/.kube/config
+    if [ ! -f "$bash_path/config" ]; then
+        echo "$bash_path/config no such file"
+    else 
+        cp -i $bash_path/config $HOME/.kube/config
+        chown $(id -u):$(id -g) $HOME/.kube/config
+    fi
     export KUBECONFIG=$HOME/.kube/config
     echo -e "---------------------\033[31m "`hostname`"just cluster success \033[0m---------------------"
 }
@@ -69,6 +73,13 @@ main() {
     rootssh_trust
     # 加入集群
     join_cluster
+
+    # 节点打标签
+    if [[ $node_type == "nvidia_gpu" ]];then
+        nvidia_gpu_label
+    elif [[ $node_type == "huawei_a910" ]];then
+        huawei_a910_label
+    fi
 }
 
 main
